refactor: Clean codes
